json受け取るようにした
